+++ packages/react-native-sdk/src/components/Participant/ParticipantView/VideoRenderer/index.tsx
@@ -97,12 +97,12 @@ export const VideoRenderer = ({
     isParticipantVideoEnabled(participant.sessionId);
 
   useEffect(() => {
-    if (Platform.OS === 'ios' && registerIosScreenshot && canShowVideo) {
-      registerIosScreenshot(participant, trackType, viewRef);
-      return () => {
-        deregisterIosScreenshot(participant, trackType);
-      };
-    }
+    if (Platform.OS !== 'ios' || !registerIosScreenshot || !canShowVideo)
+      return;
+    registerIosScreenshot(participant, trackType, viewRef);
+    return () => {
+      deregisterIosScreenshot(participant, trackType);
+    };
   }, [
     participant,
     trackType,